+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixes description id's in DatePicker for better accessibility.
+
 ## 4.49.0 - (October 6, 2020)
 
 * Changed